Massive Revamp (Removal of Heads-up, Replacement of Fitness Function etc)
1. Replace Heads-up with Tournament Evaluation
     - Board class renamed to "Table"
     - Table contains > 2 players (Default: 8)
     - Players are ranked after Tournament by profits

2. Fitness measured by Profits instead of Win/Loss ratio
     - Simple measure of Win/Loss rate does not suffice with the variance in Texas Holdem

3. Duplicate Tables
     - Tournament is repeated N times w/ same deck arrangement but different player order
     - Helps to reduce existing variance

4. Tracking Evoluation Progress
     - Save parameters of best agents in each generation

5. Different graph types can now be logged
     - Bar Graph
     - Line Graph

6. A different set of Graphs are now tracked
     - Generation Performance (Line - Average performance of each generation)
     - Population Variance (Line - Variance of each generation)

+++ BlossomPoker/HandEvaluator.h
@@ -23,8 +23,8 @@ public:
 	HandEvaluator();
 	~HandEvaluator();
 
-	double DetermineOdds_PreflopHole(std::array<std::shared_ptr<Card>, 2> _Hole);
-	double DetermineOdds_MonteCarlo(std::array<std::shared_ptr<Card>, 2> _Hole, std::array<std::shared_ptr<Card>, 5> _Community, int _TrialsAmt);
+	float DetermineOdds_PreflopHole(std::array<std::shared_ptr<Card>, 2> _Hole);
+	float DetermineOdds_MonteCarlo(std::array<std::shared_ptr<Card>, 2> _Hole, std::array<std::shared_ptr<Card>, 5> _Community, int _TrialsAmt);
 
 	int DetermineValue_5Cards(const std::array<std::shared_ptr<Card>, 5>& _Hand);
 	int DetermineValue_7Cards(const std::array<std::shared_ptr<Card>, 7>& _Hand);
@@ -55,7 +55,7 @@ public:
 
 private:
 	int HR[32487834];
-	double PreflopOdds[1326];
+	float PreflopOdds[1326];
 
 	std::mt19937 MTGenerator;
