+++ BlossomPoker/Orchastrator.h
@@ -17,14 +17,14 @@ public:
 
 	~Orchastrator();
 
-	void InitializeStrat(std::array<double,8> _Thresholds);
+	void InitializeStrat(std::array<float,8> _Thresholds);
 	void InitializeRandomStrat();
 
 	void SetOptimalStrategy(std::shared_ptr<Strategy>& _Strategy);
 	BettingAction DetermineAction();
 	
-	void SetThreshold(unsigned int _Index, double _Value);
-	std::array<double, 8> GetThresholds();
+	void SetThreshold(unsigned int _Index, float _Value);
+	std::array<float, 8> GetThresholds();
 	std::shared_ptr<BlossomAI> GetAI() { return AI; }
 
 private:
